+++ CHANGELOG.md
@@ -5,3 +5,4 @@ Yii Dependency Injection Change Log
 -----------------------
 
 - Initial release.
+- Added Service Providers support.